move unify_kernels to a new file
